Added token barer and auth for security posts endpoints

+++ src/modules/post/post.service.ts
@@ -1,8 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { PostEntity } from './entities/post.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreatePostDto } from './dto/create-post.dto';
+import { RolsEnum } from 'src/common/enums/rols.enums';
 
 @Injectable()
 export class PostService {
@@ -30,17 +31,26 @@ export class PostService {
         return post;
     }
 
-    async updatePost(id: number, updatePostDto: CreatePostDto): Promise<PostEntity> {
+    async updatePost(id: number, updatePostDto: CreatePostDto, user: any): Promise<PostEntity> {
         const post = await this.getPostById(id);
+        if (!post) throw new NotFoundException();
+
+        if (post.authorId !== user.id) throw new ForbiddenException('No tienes permiso para editar este POST');
 
         Object.assign(post, updatePostDto);
 
         return this.postRepository.save(post);
     }
 
-    async deletePost(id: number): Promise<void> {
+    async deletePost(id: number, user: any): Promise<void> {
 
         const post = await this.getPostById(id);
+        if (!post) throw new NotFoundException();
+
+        const isOwner = post.authorId === user.id;
+        const hasPrivileges = [RolsEnum.ADMIN, RolsEnum.MODERATOR].some(role => user.roles.include(role));
+
+        if (!isOwner && !hasPrivileges) throw new ForbiddenException('No tienes permiso para eliminar este POST');
         
         await this.postRepository.remove(post);
     }
